Fix payments service db and data types

+++ payments-service/src/main/java/sys/model/request/TopUpRequest.java
@@ -1,6 +1,4 @@
 package sys.model.request;
 
-import java.math.BigDecimal;
-
-public record TopUpRequest(BigDecimal amount) {
+public record TopUpRequest(Integer amount) {
 }
